Added point light, spot light and directional light. plus small bug fixing in shader class.

+++ OpenGL/OpenGL/Shader.cpp
@@ -84,9 +84,9 @@ void Shader::SetInt(const std::string& name, int value) const
 	glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
 }
 
-void Shader::SetFloat(const std::string& name, int value) const
+void Shader::SetFloat(const std::string& name, float value) const
 {
-	glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+	glUniform1f(glGetUniformLocation(ID, name.c_str()), (float)value);
 }
 
 void Shader::SetMat3(const std::string& name, glm::mat3 value) const
